24445 change elevation and depth fields to numeric (#208)
* Change elevation and depth fields to numeric
* Decimals in tests
* update CollectingEventValidator

+++ src/main/java/ca/gc/aafc/collection/api/dto/CollectingEventDto.java
@@ -23,6 +23,8 @@ import org.javers.core.metamodel.annotation.PropertyName;
 import org.javers.core.metamodel.annotation.TypeName;
 
 import javax.annotation.Nullable;
+
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -100,10 +102,10 @@ public class CollectingEventDto {
   private String dwcCountryCode;
   private String dwcStateProvince;
   private String habitat;
-  private Integer dwcMinimumElevationInMeters;
-  private Integer dwcMinimumDepthInMeters;
-  private Integer dwcMaximumElevationInMeters;
-  private Integer dwcMaximumDepthInMeters;
+  private BigDecimal dwcMinimumElevationInMeters;
+  private BigDecimal dwcMinimumDepthInMeters;
+  private BigDecimal dwcMaximumElevationInMeters;
+  private BigDecimal dwcMaximumDepthInMeters;
   private String substrate;
   private String remarks;
